# Attached-Assets/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attendance Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  employee: "employee",
  manager: "manager",
  admin: "admin",
} as const;

export interface User {
  id: number;
  clerkId: string;
  email: string;
  name: string;
  role: UserRole;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface UserSync {
  clerkId: string;
  email: string;
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export type UserRoleUpdateRole =
  (typeof UserRoleUpdateRole)[keyof typeof UserRoleUpdateRole];

export const UserRoleUpdateRole = {
  employee: "employee",
  manager: "manager",
  admin: "admin",
} as const;

export interface UserRoleUpdate {
  role: UserRoleUpdateRole;
}

export type AttendanceRecordStatus =
  (typeof AttendanceRecordStatus)[keyof typeof AttendanceRecordStatus];

export const AttendanceRecordStatus = {
  incomplete: "incomplete",
  completed: "completed",
} as const;

export type AttendanceRecordValidationStatus =
  (typeof AttendanceRecordValidationStatus)[keyof typeof AttendanceRecordValidationStatus];

export const AttendanceRecordValidationStatus = {
  pending: "pending",
  valid: "valid",
  invalid: "invalid",
} as const;

export interface AttendanceRecord {
  id: number;
  userId: number;
  user?: User;
  punchInTime: string;
  /** @nullable */
  punchOutTime?: string | null;
  /** @nullable */
  selfieUrl?: string | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  workingHours?: number | null;
  status: AttendanceRecordStatus;
  validationStatus: AttendanceRecordValidationStatus;
  /** @nullable */
  remarks?: string | null;
  overtimeRequested?: boolean;
  createdAt: string;
}

export interface AttendancePage {
  records: AttendanceRecord[];
  total: number;
  page: number;
  limit: number;
}

export interface AttendanceOrNull {
  record: AttendanceRecord | null;
}

export interface PunchInInput {
  selfieBase64: string;
  latitude: number;
  longitude: number;
}

export interface PunchOutInput {
  latitude: number;
  longitude: number;
  /** @nullable */
  selfieBase64?: string | null;
}

export type AttendanceValidationValidationStatus =
  (typeof AttendanceValidationValidationStatus)[keyof typeof AttendanceValidationValidationStatus];

export const AttendanceValidationValidationStatus = {
  valid: "valid",
  invalid: "invalid",
} as const;

export interface AttendanceValidation {
  validationStatus: AttendanceValidationValidationStatus;
  /** @nullable */
  remarks?: string | null;
}

export type OvertimeRequestStatus =
  (typeof OvertimeRequestStatus)[keyof typeof OvertimeRequestStatus];

export const OvertimeRequestStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface OvertimeRequest {
  id: number;
  userId: number;
  user?: User;
  attendanceId: number;
  hoursRequested: number;
  /** @nullable */
  reason?: string | null;
  status: OvertimeRequestStatus;
  /** @nullable */
  reviewedBy?: number | null;
  /** @nullable */
  reviewNote?: string | null;
  createdAt: string;
}

export interface OvertimeInput {
  attendanceId: number;
  hoursRequested: number;
  /** @nullable */
  reason?: string | null;
}

export type OvertimeReviewStatus =
  (typeof OvertimeReviewStatus)[keyof typeof OvertimeReviewStatus];

export const OvertimeReviewStatus = {
  approved: "approved",
  rejected: "rejected",
} as const;

export interface OvertimeReview {
  status: OvertimeReviewStatus;
  /** @nullable */
  reviewNote?: string | null;
}

export interface DashboardStats {
  totalDays: number;
  completedDays: number;
  incompleteDays: number;
  pendingOvertimeCount: number;
  avgWorkingHours: number;
  pendingValidationCount: number;
}

export interface DailyReport {
  date: string;
  records: AttendanceRecord[];
  total: number;
}

export type ListAttendanceParams = {
  userId?: number;
  date?: string;
  startDate?: string;
  endDate?: string;
  page?: number;
  limit?: number;
};

export type ListOvertimeParams = {
  status?: ListOvertimeStatus;
  userId?: number;
};

export type ListOvertimeStatus =
  (typeof ListOvertimeStatus)[keyof typeof ListOvertimeStatus];

export const ListOvertimeStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export type GetDailyReportParams = {
  date?: string;
  userId?: number;
  page?: number;
  limit?: number;
};