# Attached-Assets/lib/api-zod/src/generated/types/attendanceRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attendance Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceRecordStatus } from "./attendanceRecordStatus";
import type { AttendanceRecordValidationStatus } from "./attendanceRecordValidationStatus";
import type { User } from "./user";

export interface AttendanceRecord {
  id: number;
  userId: number;
  user?: User;
  punchInTime: string;
  /** @nullable */
  punchOutTime?: string | null;
  /** @nullable */
  selfieUrl?: string | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  workingHours?: number | null;
  status: AttendanceRecordStatus;
  validationStatus: AttendanceRecordValidationStatus;
  /** @nullable */
  remarks?: string | null;
  overtimeRequested?: boolean;
  createdAt: string;
}