# Attached-Assets/lib/api-zod/src/generated/types/dailyReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attendance Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceRecord } from "./attendanceRecord";

export interface DailyReport {
  date: string;
  records: AttendanceRecord[];
  total: number;
}