# Attached-Assets/lib/api-zod/src/generated/types/overtimeInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attendance Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface OvertimeInput {
  attendanceId: number;
  hoursRequested: number;
  /** @nullable */
  reason?: string | null;
}