# Attached-Assets/lib/api-zod/src/generated/types/overtimeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attendance Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { OvertimeRequestStatus } from "./overtimeRequestStatus";
import type { User } from "./user";

export interface OvertimeRequest {
  id: number;
  userId: number;
  user?: User;
  attendanceId: number;
  hoursRequested: number;
  /** @nullable */
  reason?: string | null;
  status: OvertimeRequestStatus;
  /** @nullable */
  reviewedBy?: number | null;
  /** @nullable */
  reviewNote?: string | null;
  createdAt: string;
}