# Attached-Assets/lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attendance Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  clerkId: string;
  email: string;
  name: string;
  role: UserRole;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}